backend/requirements.txt
Flask
google-cloud-storage
pandas
seaborn
scipy
numpy
matplotlib
protobuf